refactor(express-session.ts): add application object to SessionData interface to store application related data refactor(applyController.ts): add postLawyers function to handle POST request for lawyers application
